build: converted to class drawers all external shapes
